requirements.txt
torch
gradio
pillow
numpy
